refactor: evaluation subpackage
