+++ src/brittle_star_project/__init__.py
@@ -2,7 +2,14 @@ from .environment.env_types import Backend, Task
 from .environment.env_config import ArenaConfig, EnvConfig, MorphologyConfig
 from .environment.factory import BrittleStarEnvFactory
 from .environment.env_wrapper import BrittleStarEnv
-from .render import simulate_policy, SimulationConfig, ControlPolicy
+from .evaluation import (
+    PolicyAgent,
+    ControlPolicy,
+    load_metadata,
+    rollout_headless,
+    rollout_viewer,
+    EpisodeResult,
+)
 
 __all__ = [
     "ArenaConfig",
@@ -12,7 +19,10 @@ __all__ = [
     "EnvConfig",
     "MorphologyConfig",
     "Task",
-    "simulate_policy",
-    "SimulationConfig",
+    "PolicyAgent",
     "ControlPolicy",
+    "load_metadata",
+    "rollout_headless",
+    "rollout_viewer",
+    "EpisodeResult",
 ]
