+++ src/brittle_star_project/evaluation/__init__.py
@@ -1,0 +1,17 @@
+from __future__ import annotations
+
+from .checkpoint import load_metadata, load_params, metadata_to_configs, TrainingConfig
+from .policy import PolicyAgent, ControlPolicy
+from .rollout import rollout_headless, rollout_viewer, EpisodeResult
+
+__all__ = [
+    "load_metadata",
+    "load_params",
+    "metadata_to_configs",
+    "TrainingConfig",
+    "PolicyAgent",
+    "ControlPolicy",
+    "rollout_headless",
+    "rollout_viewer",
+    "EpisodeResult",
+]
