--- src/brittle_star_project/render/__init__.py
@@ -1,3 +1,0 @@
-from .renderer import simulate_policy, SimulationConfig, ControlPolicy
-
-__all__ = ["simulate_policy", "SimulationConfig", "ControlPolicy"]
